rewrite amount formatting util
