+++ wormhole-connect/src/hooks/useGetTokenBalances.ts
@@ -2,14 +2,10 @@ import { useDispatch, useSelector } from 'react-redux';
 import { RootState } from 'store';
 import { TokenId } from 'sdklegacy';
 import { useEffect, useState } from 'react';
-import {
-  accessBalance,
-  Balances,
-  formatBalance,
-  updateBalances,
-} from 'store/transferInput';
+import { accessBalance, Balances, updateBalances } from 'store/transferInput';
 import config, { getWormholeContextV2 } from 'config';
 import { TokenConfig } from 'config/types';
+import { formatAmount } from 'utils/amount';
 import { chainToPlatform } from '@wormhole-foundation/sdk-base';
 import { getTokenBridgeWrappedTokenAddress } from 'utils/sdkv2';
 import { Chain, TokenAddress } from '@wormhole-foundation/sdk';
@@ -133,7 +129,7 @@ const useGetTokenBalances = (
             const balance = result[tokenAddress];
             let formatted: string | null = null;
             if (balance !== null) {
-              formatted = formatBalance(chain, tokenConfig, BigInt(balance));
+              formatted = formatAmount(chain, tokenConfig, balance);
             }
             updatedBalances[tokenConfig.key] = {
               balance: formatted,
